Add /mod, make d.moderate into modUtils, merge watchUser, modify command index to ignore directories
