+++ src/discord/commands/guild/m.report.ts
@@ -7,7 +7,7 @@ import {
 } from 'discord-api-types/v10';
 import { MessageCommand } from '../../@types/commandDef';
 import commandContext from '../../utils/context';
-import { modResponse } from './d.moderate';
+import { modResponse } from '../../utils/modUtils';
 
 const F = f(__filename);
